Improve error messages for OAuth
