+++ .changeset/quiet-snakes-marry.md
@@ -1,0 +1,7 @@
+---
+'eurosky-portal': patch
+---
+
+Improve OAuth error handling
+
+We now display a flash message above the form an OAuth error occurs after redirecting to the OAuth provider (assuming they redirect back to us). The other error messages have also been improved.
